# api/manager/swagger.yaml
basePath: /
definitions:
...
  d7y_io_dragonfly_v2_manager_models.Oauth:
    properties:
      bio:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      name:
        type: string
      redirect_url:
        type: string
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.CreateClusterRequest:
    properties:
      bio:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterClientConfig'
      scheduler_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterConfig'
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterScopes'
      seed_peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
    required:
    - name
    - peer_cluster_config
    - scheduler_cluster_config
    - seed_peer_cluster_config
    type: object
  d7y_io_dragonfly_v2_manager_types.CreateClusterResponse:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterClientConfig'
      scheduler_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterConfig'
      scheduler_cluster_id:
        type: integer
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterScopes'
      seed_peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
      seed_peer_cluster_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.CreateOauthRequest:
    properties:
      bio:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      name:
        enum:
        - github
        - google
        type: string
      redirect_url:
        type: string
    required:
    - client_id
    - client_secret
    - name
    type: object
...
  d7y_io_dragonfly_v2_manager_types.GetClusterResponse:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterClientConfig'
      scheduler_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterConfig'
      scheduler_cluster_id:
        type: integer
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterScopes'
      seed_peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
      seed_peer_cluster_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.SchedulerClusterClientConfig:
    properties:
      load_limit:
        maximum: 2000
        minimum: 1
        type: integer
    type: object
  d7y_io_dragonfly_v2_manager_types.SchedulerClusterConfig:
    properties:
      candidate_parent_limit:
        maximum: 20
        minimum: 1
        type: integer
      filter_parent_limit:
        maximum: 1000
        minimum: 10
        type: integer
      job_rate_limit:
        maximum: 1000000
        minimum: 1
        type: integer
    type: object
  d7y_io_dragonfly_v2_manager_types.SchedulerClusterScopes:
    properties:
      cidrs:
        items:
          type: string
        type: array
      hostnames:
        items:
          type: string
        type: array
      idc:
        type: string
      location:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig:
    properties:
      load_limit:
        maximum: 50000
        minimum: 1
        type: integer
    type: object
...
  d7y_io_dragonfly_v2_manager_types.UpdateClusterRequest:
    properties:
      bio:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterClientConfig'
      scheduler_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterConfig'
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterScopes'
      seed_peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
    type: object
  d7y_io_dragonfly_v2_manager_types.UpdateClusterResponse:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterClientConfig'
      scheduler_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterConfig'
      scheduler_cluster_id:
        type: integer
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerClusterScopes'
      seed_peer_cluster_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
      seed_peer_cluster_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.UpdateOauthRequest:
    properties:
      bio:
        type: string
      client_id:
        type: string
      client_secret:
        type: string
      name:
        enum:
        - github
        - google
        type: string
      redirect_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    url: https://d7y.io
  description: Dragonfly Manager Server
  license:
    name: Apache 2.0
  title: Dragonfly Manager
  version: 1.0.0
paths:
...
  /api/v1/clusters:
    get:
      consumes:
      - application/json
      description: Get Clusters
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.GetClusterResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Clusters
      tags:
      - Cluster
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: Cluster
        in: body
        name: Cluster
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateClusterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create Cluster
      tags:
      - Cluster
  /api/v1/clusters/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy Cluster
      tags:
      - Cluster
    get:
      consumes:
      - application/json
      description: Get Cluster by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.GetClusterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Cluster
      tags:
      - Cluster
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Cluster
        in: body
        name: Cluster
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateClusterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update Cluster
      tags:
      - Cluster
...
  /api/v1/oauth:
    get:
      consumes:
      - application/json
      description: Get Oauths
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Oauth'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Oauths
      tags:
      - Oauth
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: Oauth
        in: body
        name: Oauth
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateOauthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Oauth'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create Oauth
      tags:
      - Oauth
  /api/v1/oauth/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy Oauth
      tags:
      - Oauth
    get:
      consumes:
      - application/json
      description: Get Oauth by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Oauth'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Oauth
      tags:
      - Oauth
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Oauth
        in: body
        name: Oauth
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateOauthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Oauth'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update Oauth
      tags:
      - Oauth
...
  /oapi/v1/clusters:
    get:
      consumes:
      - application/json
      description: Get Clusters
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.GetClusterResponse'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Clusters
      tags:
      - Cluster
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: Cluster
        in: body
        name: Cluster
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateClusterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create Cluster
      tags:
      - Cluster
  /oapi/v1/clusters/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy Cluster
      tags:
      - Cluster
    get:
      consumes:
      - application/json
      description: Get Cluster by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.GetClusterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Cluster
      tags:
      - Cluster
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Cluster
        in: body
        name: Cluster
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateClusterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update Cluster
      tags:
      - Cluster
...
swagger: "2.0"
...